Include/auth bug #3 fixed

+++ datadinascita/birthdays/views.py
@@ -1,9 +1,8 @@
-import logging
 import csv
 
 from google.appengine.ext import blobstore
 
-from django.http import  HttpResponse
+from django.http import  HttpResponse, HttpResponseRedirect
 
 from datadinascita.birthdays.utils import *
 from datadinascita.birthdays.forms import AddForm
@@ -12,19 +11,24 @@ def index(request):
     return render_to_response('index.html', {'auth_url': get_auth_url(request.META['PATH_INFO'])})
 
 def fof(request, query):
-	"""404 pages"""
-	return render_to_response('404.html', {'query': query})
+    """404 pages"""
+    return render_to_response('404.html', {'query': query})
 
 def search(request):
     return render_to_response('search.html', {})
 
 def list(request):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     people = modify_people(Person.all().filter("owner =", users.get_current_user()))
-    return render_to_response('list.html', {'people': people, 'count': len(people), 'auth_url': get_auth_url(request.META['PATH_INFO'])})
+    return render_to_response('list.html', {'people': people, 'count': len(people),
+                                            'auth_url': get_auth_url(request.META['PATH_INFO'])})
 
 def export(request):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     people = modify_people(Person.all().filter("owner =", users.get_current_user()))
     response = HttpResponse(mimetype='text/csv')
     response['Content-Disposition'] = 'attachment; filename=birthdays.csv'
@@ -38,7 +42,9 @@ def export(request):
 
 
 def csv_upload(request):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     try:
         upload_url = blobstore.create_upload_url('/upload_csv/')
     except:
@@ -46,11 +52,15 @@ def csv_upload(request):
     return render_to_response('import.html', {'upload_url': upload_url})
 
 def import_failed(request):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     return render_to_response('import_failed.html', {})
 
 def erase(request):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     people = Person.all().filter("owner =", users.get_current_user())
     for person in people:
         person.delete()
@@ -58,7 +68,9 @@ def erase(request):
     return HttpResponseRedirect('/people')
 
 def add(request):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     if request.method == 'POST':
         form = AddForm(request.POST)
 
@@ -84,6 +96,8 @@ def add(request):
     return show_new_person(form, request)
 
 def edit(request, id):
-    check_auth(request)
+    if not users.get_current_user():
+        return HttpResponseRedirect(users.create_login_url(request.META['PATH_INFO']))
+
     return render_to_response('index.html', {'query': id})
